requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
ffmpeg-python==0.2.0
soundfile==0.12.1
numpy==1.25.2
openai-whisper==20231117
fasttext==0.9.2
python-multipart==0.0.6
pytest==7.4.0 